fix: clear all clang-tidy findings across engine, editor, overlay, and tests
Twenty-two findings under the .clang-tidy config (bugprone + analyzer +
performance, WarningsAsErrors='*') across 10 translation units. All
behaviour-preserving: explicit static_casts, hoisted optional derefs
past guards, restructured loops to never re-enter fread/fgets on an
EOF stream, and an else-if + FCITX_ASSERT pair for an unproven branch
invariant.

Engine (addon/src/schnelle-umlaute.cpp):
- Mark SchnelleUmlauteEngine final — the class is a leaf addon never
  subclassed (only instantiated by SchnelleUmlauteEngineFactory).
  Silences the VirtualCall finding on reloadConfig() in the
  constructor, with the tiny bonus of allowing de-virtualization.
- commitCyclingValue / scheduleTimeout: hoist *state->cyclingInput_
  and *state->waitingKey_ derefs to immediately after the optional
  guard so bugprone-unchecked-optional-access can prove safety across
  the intervening cancelTimeout() / getEffectiveDelay() calls.
- Three narrowing conversions wrapped in static_cast: setCursor(int)
  vs textLength() (size_t), overlayShow(int) vs cyclingIndex_
  (size_t), and the ASCII tolower arithmetic in parseFirstChar.

Other production sources:
- hand_classifier.cpp: same ASCII tolower static_cast<char> fix as
  parseFirstChar.
- editor/SettingsModel.cpp + overlay/main.cpp: QString::indexOf
  returns qsizetype (long long); cast to int at three call sites
  where the offsets are bounded by line length.

Shared header (addon/mappings-io.h):
- parseMappings(): clang-analyzer-unix.Stream flagged the fgets/fgetc
  EOF-boundary branch as a re-entry into an EOF stream. Add a
  streamEnded flag and an explicit EOF check in the drain-truncated-
  line loop so fgets is never called on a stream past EOF.

Tests (tests/*.cpp + tests/testdir.h.in):
- testdir.h.in: rename the _TEST_TESTDIR_H_ include guard, which used
  the reserved leading-underscore + uppercase pattern.
- testschnelleumlaute.cpp four findings: collapse branch-clone in
  countDualLeaderWithinWordCollisions; add explicit else-if +
  FCITX_ASSERT for the lastLetter range in typeTextWordBoundaryOverlap
  so kLetterCodes is never indexed out of range; static_cast<char>
  for the ASCII tolower in Test 57; reset the global testDispatcher
  to nullptr before main returns so the dispatcher's stack address
  no longer outlives its scope.
- testmappingsio.cpp: rewind() has no error reporting; switch to
  fseek + explicit failure check.
- testmappinglistmodelio.cpp + testsettingsmodel.cpp: rewrite the
  fread loop with the same streamEnded pattern used in
  mappings-io.h, so fread is never called on an EOF stream.

Local checks all green: clang-tidy on 20 translation units (5
production + 4 editor + 2 overlay + 9 test) reports 0 findings;
ctest 9/9 in ~9.02s under the new flags.

+++ addon/overlay/main.cpp
@@ -41,7 +41,7 @@ QString loadInitialTheme() {
             continue;
         }
         if (section != QLatin1String("Theme")) continue;
-        const int eq = line.indexOf('=');
+        const int eq = static_cast<int>(line.indexOf('='));
         if (eq < 0) continue;
         if (line.left(eq) == QLatin1String("Theme")) {
             return line.mid(eq + 1).trimmed();
